fix a regex for a line comment
The previous regular expression for a line comment requires that //
should be immediately followed by a character other than the backslash
or newline character. So it did not allow 1) a line comment whose whole
line just consists of // or 2) a line comment where // is immediately
followed by the backslash character.